Add "last" setter to List.
This allows assigning directly as `list.last = value` instead of
having to use the longer `list[list.length - 1] = value`.
It's just an abbreviation, like `list.last` is for reading.

If you use a List as a stack (which is a perfectly good use of it),
modifying the last element like this is a natural operation.

R=sgjesse@google.com

Review URL: https://codereview.chromium.org//315173005

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart/pkg/third_party/html5lib@37255 260f80e4-7a28-3924-810f-c04153c831b5