fix parse error with foreignObject end tag
it appears we inherited this from html5lib:
https://github.com/html5lib/html5lib-python/blob/5e3d432998b027cd2d200446f157d13a3461a92b/html5lib/html5parser.py#L2448
You'll notice the node.name.translate(asciiUpper2Lower) immediately below, in the same comparison

Most likely, this is not caught because it's not common to display HTML parse errors, and the actual tree construction is correct.

This fix does match the spec:
https://html.spec.whatwg.org/multipage/syntax.html#parsing-main-inforeign
See "Any other end tag, step 2":
> If node's tag name, converted to ASCII lowercase, is not the same as the tag name of the token, then this is a parse error.

We were missing the converted to lowercase step.

R=sigmund@google.com

Review URL: https://codereview.chromium.org//1222713008.