Fix infinite loop with bad button close tag
Closes #122

Add 'button' to the tags that are processed with `endTagBlock`. The spec
at https://html.spec.whatwg.org/multipage/parsing.html includes 'button'
in the list of end tags with this treatement.

Add a new `.dat` file for regression test. It include an `#info` field
which is new but can be parsed safety to mark the issue for which it is
a test. This test loops endlessly before the fix and passes after the
fix.

Note that since errors are currently untested and test fail if they are
tested the error will be left without content for now.